PO-150: Remove unnecessary interfaces
